python_agent/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
langchain>=0.2.6
langchain-core>=0.2.22
langchain-google-genai>=1.0.7
langgraph>=0.1.14
pinecone-client==3.2.2
cohere==5.5.6
tavily-python==0.3.3
httpx==0.27.0
pydantic==2.7.4
numpy==1.26.4